Adding tests to cover social button components
