+++ src/components/__tests__/now/playing.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import Playing from '../../now/playing';
+
+const testGame = {
+    steamId: '220',
+    name: 'Half Life 2'
+}
+
+describe('Playing', () => {
+  it('renders correctly', () => {
+    const tree = renderer
+      .create(<Playing game={testGame}/>)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
