+++ src/components/__tests__/socialButtons/emailSocialButton.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import EmailSocialButton from '../../socialButtons/emailSocialButton';
+
+describe('Email Social Button', () => {
+  it('renders correctly', () => {
+    const tree = renderer
+      .create(<EmailSocialButton />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
